backend/pyproject.toml
[tool.poetry]
name = "hack-change-backend"
version = "0.1.0"
description = ""
authors = ["allelleo <dev.allelleo@internet.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
tortoise-orm = "^0.22.1"
httpx = "^0.27.2"
pyjwt = "^2.10.1"
aerich = "^0.7.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
black = "^24.10.0"
colorlog = "^6.9.0"
pydantic = {extras = ["email"], version = "^2.10.2"}
prometheus-fastapi-instrumentator = "^7.0.0"
python-multipart = "^0.0.18"




[tool.aerich]
tortoise_orm = "hack_change_backend.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
